+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added support for the package manager uv. ([#362](https://github.com/heroku/buildpacks-python/pull/362))
+
 ## [1.0.0] - 2025-05-02
 
 ### Removed